Restructured code files, implemented register
