+++ Assets/Prefabs/SensapexLink.prefab
@@ -9,7 +9,7 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 7137387518804267671}
-  - component: {fileID: 9045126198136274274}
+  - component: {fileID: 6854367561162462583}
   m_Layer: 0
   m_Name: SensapexLink
   m_TagString: Untagged
@@ -31,7 +31,7 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &9045126198136274274
+--- !u!114 &6854367561162462583
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -43,5 +43,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 1d1472904c7237a449b283ca843dc7a2, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  _serverIp: localhost
-  _serverPort: 8080
+  serverIp: localhost
+  serverPort: 8080
+  calibrateOnConnect: 0
