+++ Assets/Scripts/SensapexLink.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ae8d2941f715a224292917463bbd6ce1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
