+++ Assets/Scripts/SensapexLink/CommunicationManager.cs.meta
