+++ Assets/Scripts/SensapexLink/DataFormats.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: afe129e87bfbee841a8a9bb145c658fa
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
